requirements.txt
asgiref==3.5.2
asttokens==2.2.1
attrs==21.4.0
backcall==0.2.0
black==23.1.0
click==8.1.3
decorator==5.1.1
Django==4.0.5
djangorestframework==3.13.1
executing==1.2.0
importlib-resources==5.8.0
inflection==0.5.1
ipdb==0.13.11
ipython==8.10.0
jedi==0.18.2
jsonschema==4.6.0
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
packaging==23.0
parso==0.8.3
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.0.0
prompt-toolkit==3.0.37
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.6.2
traitlets==5.9.0
uritemplate==4.1.1
wcwidth==0.2.6
zipp==3.8.0
